Move function in resource.
